requirements.txt
streamlit
matplotlib.pyplot
sympy